Write updates to session stores at the end of the request
The ->finalize method is called when the headers are sent to the client;
at this time, the store is saved, and the state is sent to the client.
However, this is insufficient for streaming applications, where further
changes to the session store may be made later in the response.

Use psgix.cleanup.handlers (if available), or a DESTROY block in the
$env if not, to write the session store at the end of the request.