Move late session storage to an option, to prevent race conditions
Storing the session after the request has completed does open the
session to a potential race condition, if the client successfully makes
a request before the session has been saved.

Turn the late storage behavior into an option, defaulting to off.  This
makes the potential error _consistent_ (late alterations to the session
are missed) rather than a race condition; the failure mode is thus
easier to diagnose and reason about.

Fixes stevan/plack-middleware-session#29