Ensure that change_id is updated before storage
df17b414 moved the change_id behavior into its own method; during the
extraction, it moved the behavior to after the session had been
committed.  This is incorrect, as it caused the store to be updated
under the old id, then immediately removed and never re-added.

This was hidden by the succeeding commit, a9aba010, which moved all
session storage being delayed until the end of the request.  Now that
that is not the default, the proper order of change_id then commit must
be restored.